+++ bin/dev-server.sh
@@ -12,4 +12,6 @@ export PYTHONWARNINGS=always
 #export CODEX_THROTTLE_OPDS=10
 #export CODEX_THROTTLE_USER=10
 #uv run python3 -X tracemalloc ./codex/run.py
+export DJANGO_SETTINGS_MODULE=codex.settings
+#uv run righttyper --all-files --overwrite codex/run.py
 uv run python3 ./codex/run.py